Initial working PlantUml encoding algorithm
This code is a very literal port of the php/javascript example code
found in the PlantUML documentation. I haven't made any attempts to
follow PowerShell idioms or otherwise improve the code.

It's basically a proof-of-concept, to show that PowerShell can produce
an encoded string that matches the output of the reference
implementation.